find-by-key function was written
I checked the test cases and data structure. We need to return value for id given. in the sample file structure is array. and each element for array object. we can use find method for each element then once we find element we can check id property of element(object)